Title: 합과 차의 나눗셈, Time: 131ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240802/합과 차의 나눗셈/divide-of-sum-and-sub.py
@@ -1,2 +1,2 @@
-a,b= map(int,input().split())
-print(round(float((a+b)/(a-b)),2))
+a,b= map(float,input().split())
+print(round((a+b)/(a-b),2))
